# Don't enable execution tracer tasks if tracer is not enabled