@rakyll 2018-04-23
@ramonza, it is not doing it right now because user end up turning on the execution tracer at any time. This may result in spans without parents and tasks.

For us, it makes sense no check this when we are starting a OpenCensus associated task. A distributed span is the higher level logical group.